修复:LamaMPEInpainter中的数据类型转换以实现一致的图像处理

+++ manga_translator/inpainting/inpainting_lama_mpe.py
@@ -107,9 +107,9 @@ class LamaMPEInpainter(OfflineInpainter):
                     img_inpainted_torch = self.model(img_torch, mask_torch)
 
         if isinstance(self.model, LamaFourier):
-            img_inpainted = (img_inpainted_torch.cpu().squeeze_(0).permute(1, 2, 0).numpy() * 255.).astype(np.uint8)
+            img_inpainted = (img_inpainted_torch.cpu().squeeze_(0).permute(1, 2, 0).float().numpy() * 255.).astype(np.uint8)
         else:
-            img_inpainted = ((img_inpainted_torch.cpu().squeeze_(0).permute(1, 2, 0).numpy() + 1.0) * 127.5).astype(np.uint8)
+            img_inpainted = ((img_inpainted_torch.cpu().squeeze_(0).permute(1, 2, 0).float().numpy() + 1.0) * 127.5).astype(np.uint8)
         if new_h != height or new_w != width:
             img_inpainted = cv2.resize(img_inpainted, (width, height), interpolation = cv2.INTER_LINEAR)
         ans = img_inpainted * mask_original + img_original * (1 - mask_original)
